+++ test/setup.ts
@@ -8,5 +8,5 @@ import { env } from "./env";
 export const test = createViemTest(mainnet, {
   forkUrl: env().MAINNET_RPC_URL,
   chainId: mainnet.id,
-  forkBlockNumber: 23891510n,
+  forkBlockNumber: 24382381n,
 });